# perf: Optimize cutoff date range APIs - 99% faster ⚡