--- a/backend/src/modules/hr/timekeeping/employee-timekeeping/employee-timekeeping.controller.ts
+++ b/backend/src/modules/hr/timekeeping/employee-timekeeping/employee-timekeeping.controller.ts
@@ -167,6 +167,14 @@ export class EmployeeTimekeepingController {
     );
   }
 
+  @Get('cutoff-date-range-lite')
+  async employeeTimekeepingCutoffDateRangeLite(@Res() response: Response) {
+    return this.utilityService.responseHandler(
+      this.employeeTimekeepingService.getEmployeeTimekeepingCutoffDateRangeLite(),
+      response,
+    );
+  }
+
   @Get('employee')
   async employeeTimekeepingInfo(
     @Query() params: EmployeeTimekeepingDTO,